Move admin components to new folder
