+++ components/admin/medical-details.tsx
