+++ components/admin/sensitive-data.tsx
