+++ pages/applications/reports/index.tsx
@@ -1,12 +1,12 @@
 import { GetServerSideProps, GetServerSidePropsResult } from 'next';
 import React from 'react';
-import Sidebar from '../../../components/applications/sidebar';
-import { ButtonLink } from '../../../components/button';
-import { HeadingOne } from '../../../components/content/headings';
-import Layout from '../../../components/layout/staff-layout';
 import { HackneyGoogleUser } from '../../../domain/HackneyGoogleUser';
 import { UserContext } from '../../../lib/contexts/user-context';
 import { getAuth, getSession } from '../../../lib/utils/auth';
+import Layout from '../../../components/layout/staff-layout';
+import Sidebar from '../../../components/admin/sidebar';
+import { ButtonLink } from '../../../components/button';
+import { HeadingOne } from '../../../components/content/headings';
 
 interface ReportsProps {
   user: HackneyGoogleUser;
